requirements.txt
fastapi
uvicorn
python-multipart
pytesseract
Pillow
opencv-python-headless
python-dateutil
pydantic
pdf2image
pymupdf
python-dotenv
mysql-connector-python